refactor: move constants and HTTP response function to a separate file
