+++ handlers/deleteUserHandler.ts
@@ -1,20 +1,7 @@
 import { IncomingMessage, OutgoingHttpHeaders, ServerResponse } from "http";
 import { validate, v4 as uuidv4 } from "uuid";
 import { User, users } from "../controllers/userController";
-
-const CONTENT_TYPE: OutgoingHttpHeaders = {
-  "Content-Type": "application/json",
-};
-export const sendResponse = (
-  res: ServerResponse,
-  code: number,
-  contentType: OutgoingHttpHeaders,
-  data: object | undefined,
-): void => {
-  const jsonData: string = JSON.stringify(data);
-  res.writeHead(code, contentType);
-  res.end(jsonData);
-};
+import { CONTENT_TYPE, sendResponse } from "../helpers/helpers";
 
 export const deleteUserHandler = (res: ServerResponse, url: string): void => {
   const id: string = url?.split("/").slice(-1)[0];
